feat(kernel-line): replace the origin line method with adaptive-line method
In adaptive strategy, adaptive-line method will be used now. In strategy picker, the origin line
will also be replace by the new adaptive-line method. In adaptive-line method, it can switch
betweeen line method and vector-row method (it can switch vector-row method if the LDS size is not
enough for running line method). But in pure line method, it may cause an memory error.