fix(kernel-line): fix incorrect data storing logic of vector-row method in adaptive line
In adaptive line, if vector-row method is selected, the results can be wrong when the rows for a
block is not a multiple of vector number in a block. This is caused by the incorrect condition of
"which threads can store data back". In this fix, we use variable `ACTIVE_VECTORS` to control the
condition.