fix(suitesparse-dl): fix incorrect "NNZ" values stored into binary csr file header
Before this commit, if the matrix is symmetric, the "NNZ" value that is written into the binary
csr file can be incorrect. Because we use the non-zeros number in matrix market file header,
which is not the real matrix NNZ when the matrix is symmetric.
This commit fixed above bug.